run just integration using ingest
